server/src/database/mysql.py
import os
import mysql.connector
from dotenv import load_dotenv
load_dotenv()

mydb = mysql.connector.connect(
        host=os.getenv("DB_HOST"),
        port=os.getenv("DB_PORT"),
        user=os.getenv("DB_USER"),
        password=os.getenv("DB_PWD"),
        database=os.getenv("DB_NAME"),
)

mycursor = mydb.cursor(dictionary=True)
